# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Game On Tabletop API
  description: REST API for Game On Tabletop — a platform for organising tabletop
    game nights, tracking game collections, scheduling events, and managing group
    RSVPs.
  version: 1.0.0
paths:
...
  /api/users/me/collection:
    get:
      tags:
      - collection
      summary: List the authenticated user's game collection
      operationId: list_my_collection_api_users_me_collection_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectionEntryResponse'
                title: Response List My Collection Api Users Me Collection Get
        '401':
          description: Unauthorized
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - collection
      summary: Add a game to the authenticated user's collection
      operationId: add_to_collection_api_users_me_collection_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionEntryCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionEntryResponse'
        '401':
          description: Unauthorized
        '409':
          description: Game with this BGG ID already in collection
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/me/collection/{entry_id}:
    patch:
      tags:
      - collection
      summary: Update a collection entry
      operationId: update_collection_entry_api_users_me_collection__entry_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: entry_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Entry Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionEntryUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionEntryResponse'
        '401':
          description: Unauthorized
        '403':
          description: Not your collection entry
        '404':
          description: Collection entry not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - collection
      summary: Remove a game from the authenticated user's collection
      operationId: delete_collection_entry_api_users_me_collection__entry_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: entry_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Entry Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: Unauthorized
        '403':
          description: Not your collection entry
        '404':
          description: Collection entry not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{user_id}/profile:
    get:
      tags:
      - collection
      summary: Get a user's public profile by ID
      operationId: get_user_profile_api_users__user_id__profile_get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublicProfileResponse'
        '404':
          description: User not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{user_id}/collection:
    get:
      tags:
      - collection
      summary: Get a user's game collection (visibility-filtered)
      operationId: get_user_collection_api_users__user_id__collection_get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectionEntryResponse'
                title: Response Get User Collection Api Users  User Id  Collection
                  Get
        '401':
          description: Unauthorized
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CollectionEntryCreate:
      properties:
        game_title:
          type: string
          title: Game Title
        bgg_game_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Bgg Game Id
        game_thumbnail_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Game Thumbnail Url
        min_players:
          anyOf:
          - type: integer
          - type: 'null'
          title: Min Players
        max_players:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Players
        complexity:
          anyOf:
          - type: number
          - type: 'null'
          title: Complexity
        status:
          type: string
          enum:
          - own
          - wishlist
          - want_to_play
          - previously_owned
          title: Status
          default: own
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        acquired_at:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Acquired At
        collection_visible_to:
          type: string
          enum:
          - public
          - friends
          - private
          title: Collection Visible To
          default: friends
      type: object
      required:
      - game_title
      title: CollectionEntryCreate
    CollectionEntryResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        bgg_game_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Bgg Game Id
        game_title:
          type: string
          title: Game Title
        game_thumbnail_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Game Thumbnail Url
        min_players:
          anyOf:
          - type: integer
          - type: 'null'
          title: Min Players
        max_players:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Players
        complexity:
          anyOf:
          - type: number
          - type: 'null'
          title: Complexity
        status:
          type: string
          title: Status
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        source:
          type: string
          title: Source
        acquired_at:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Acquired At
        collection_visible_to:
          type: string
          title: Collection Visible To
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - bgg_game_id
      - game_title
      - game_thumbnail_url
      - min_players
      - max_players
      - complexity
      - status
      - notes
      - source
      - acquired_at
      - collection_visible_to
      - created_at
      - updated_at
      title: CollectionEntryResponse
    CollectionEntryUpdate:
      properties:
        status:
          anyOf:
          - type: string
            enum:
            - own
            - wishlist
            - want_to_play
            - previously_owned
          - type: 'null'
          title: Status
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        collection_visible_to:
          anyOf:
          - type: string
            enum:
            - public
            - friends
            - private
          - type: 'null'
          title: Collection Visible To
        acquired_at:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Acquired At
      type: object
      title: CollectionEntryUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UserPublicProfileResponse:
      properties:
        id:
          type: string
          title: Id
        username:
          type: string
          title: Username
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        avatar_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar Url
      type: object
      required:
      - id
      - username
      - display_name
      - avatar_url
      title: UserPublicProfileResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer